Extract shared openPostDetailStats helper into ActivityNavigator
Collapses the duplicated post-tap navigation in NewStatsActivity and
MostViewedDetailActivity into one ActivityNavigator method, and moves the
legacy statsDetailItemType mapping out of the new-stats UI-state layer to
the navigation boundary.
