+++ WordPress/src/main/java/org/wordpress/android/modules/ApplicationModule.java
@@ -21,6 +21,7 @@ import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadStep;
 import org.wordpress.android.ui.jetpack.backup.download.BackupDownloadStepsProvider;
 import org.wordpress.android.ui.jetpack.restore.RestoreStep;
 import org.wordpress.android.ui.jetpack.restore.RestoreStepsProvider;
+import org.wordpress.android.ui.mysite.SelectedSiteRepository;
 import org.wordpress.android.ui.sitecreation.SiteCreationStep;
 import org.wordpress.android.ui.sitecreation.SiteCreationStepsProvider;
 import org.wordpress.android.util.BuildConfigWrapper;
@@ -118,8 +119,8 @@ public abstract class ApplicationModule {
     }
 
     @Provides
-    public static ActivityNavigator provideActivityNavigator(@ApplicationContext Context context) {
-        return new ActivityNavigator();
+    public static ActivityNavigator provideActivityNavigator(SelectedSiteRepository selectedSiteRepository) {
+        return new ActivityNavigator(selectedSiteRepository);
     }
 
     @Provides
