+++ WordPress/src/test/java/org/wordpress/android/ui/StatsDetailItemTypeTest.kt
@@ -1,10 +1,10 @@
-package org.wordpress.android.ui.newstats.mostviewed
+package org.wordpress.android.ui
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import org.wordpress.android.ui.stats.StatsConstants
 
-class MostViewedCardUiStateTest {
+class StatsDetailItemTypeTest {
     @Test
     fun `post type maps to post detail item type`() {
         assertThat(statsDetailItemType("post")).isEqualTo(StatsConstants.ITEM_TYPE_POST)
